Run integration tests in CI
