+++ .env.integration.example
@@ -1,0 +1,1 @@
+OPEN_ROUTER_API_KEY=sk-or-v1-...
