+++ package.json
@@ -221,16 +221,16 @@
 		"build:webview": "cd webview-ui && npm run build",
 		"changeset": "changeset",
 		"check-types": "tsc --noEmit",
-		"compile": "npm run check-types && npm run lint && node esbuild.js",
-		"compile-tests": "tsc -p . --outDir out",
+		"compile": "tsc -p . --outDir out && node esbuild.js",
+		"compile:integration": "tsc -p tsconfig.integration.json",
 		"install:all": "npm install && cd webview-ui && npm install",
 		"lint": "eslint src --ext ts && npm run lint --prefix webview-ui",
 		"package": "npm run build:webview && npm run check-types && npm run lint && node esbuild.js --production",
-		"pretest": "npm run compile-tests && npm run compile && npm run lint",
+		"pretest": "npm run compile && npm run compile:integration",
 		"dev": "cd webview-ui && npm run dev",
 		"test": "jest && npm run test:webview",
 		"test:webview": "cd webview-ui && npm run test",
-		"test:extension": "vscode-test",
+		"test:extension": "npm run build && npm run compile:integration && npx dotenvx run -f .env.integration -- vscode-test",
 		"prepare": "husky",
 		"publish:marketplace": "vsce publish && ovsx publish",
 		"publish": "npm run build && changeset publish && npm install --package-lock-only",
@@ -245,6 +245,7 @@
 	"devDependencies": {
 		"@changesets/cli": "^2.27.10",
 		"@changesets/types": "^6.0.0",
+		"@dotenvx/dotenvx": "^1.34.0",
 		"@types/diff": "^5.2.1",
 		"@types/diff-match-patch": "^1.0.36",
 		"@types/jest": "^29.5.14",
@@ -255,7 +256,6 @@
 		"@typescript-eslint/parser": "^7.11.0",
 		"@vscode/test-cli": "^0.0.9",
 		"@vscode/test-electron": "^2.4.0",
-		"dotenv": "^16.4.7",
 		"esbuild": "^0.24.0",
 		"eslint": "^8.57.0",
 		"husky": "^9.1.7",
